remove inline for tests, may be we can just use interfaces
