+++ build.gradle
@@ -59,7 +59,6 @@ dependencies {
     testImplementation 'org.assertj:assertj-core:3.8.0'
     testImplementation("org.junit.platform:junit-platform-runner:1.0.0")
     testImplementation 'org.mockito:mockito-core:2.7.22'
-    testImplementation 'org.mockito:mockito-inline:2.7.22'
 }
 
 compileKotlin {
