--- src/test/resources/mockito-extensions/org.mockito.plugins.MockMacker
@@ -1,1 +1,0 @@
-mock-maker-inline
